# lib/api-zod/src/generated/types/ticketAttachmentDetail.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TicketAttachmentDetail {
  id: number;
  ticketId: number;
  filename: string;
  contentType: string;
  size: number;
  data: string;
  createdAt: Date;
}